+++ vscode/CHANGELOG.md
@@ -2,7 +2,20 @@
 
 ## [Upcoming release]
 Various bugfixes reported @ github
-A bug where the web view shows "null" at the top of the page
+
+## [0.2.4]
+Added new logic for write*Metric snippets
+Added tooltips for write*Metric snippets
+Added support for command-runner integration
+Added support for rule-runner integration
+Various bugfixes/improvements
+
+## [0.2.1]
+Finishing touches on the split scripts
+New web view, it no longer shows null at the top
+
+## [0.2.0]
+Added support for the new script map structure
 
 ## [0.1.6]
 Added ignore functionality to comments